Added the help command! Bot triggers on @ mention!
This is a big advancment in the bots abilities! Now you can use the `!help` command to get information on all the server commands. This is even more useful because command creators MUST have a help piece attatched to their Command struct, this command just draws from that.

The bot can also be triggered by @'ing it, to prevent prefix conflicts with other bots.

===

Reworks:
- The message content has the prefix removed from the message before passing it to the commands themselves.
- Discord error reports are now streamlined to a single function in the foundation file
- Minor text fixes within the help pages
- Verbose logging from the info file has been removed for clenliness
- Probably some really wild thing I'm forgetting

===
Known issues:
- the info command does not load in version data properly from the json config file. Dunno why but it does.

===
Where to go from here:
- Built-in and configurable low level moderation features to censor rude words, manage users, and log messages
- Add an API for interacting with users via emojis on text blocks
- Figure out how to DM things to people.